Removed Forecast from Router and Created a service
