+++ services/index.js
@@ -1,0 +1,3 @@
+var Forecast = require('./forecastService')
+
+module.exports = Forecast;
